First rev of the full schematic and layout (very rough)
My traces are shit bc I don't know how to draw traces properly.
I didn't do the diode traces because I was struggling with vias.
The diodes should probably be moved bc I wasn't sure where was best to place them for ideal trace drawing.
I think we should have multiple traces of vsolar and gnd that cross each of the bends.
Multiple traces instead of one large one to prevent crack propagation. Better yet: multiple large ones.
Any copper in the bends should run perpendicular to the bending axis per manufacturer recommendations.
I made through hole pads for the mounting holes, but didn't place them in the right spots in the layout. The mounting holes are not at nice intervals in the CAD for good reason, but the kicad grid snap wasn't cutting it.
The cells are also not precisely placed. They are like +-2 mm or so, but they don't all line up perfectly which must change, for I am seriously irked by it. Now I know why nobody's board holes line up properly: they just put them in random spots near where they should be :skull:. Is there a way to position components by offsetting it from things by a float value (like dimension tool in CAD)?
Each panel has 8 cells in series. All panels are in parallel. We might find out from testing that we need to remove another cell from the body-fixed panel for the camera to freely glide out. In this case, we can put 9 cells on another panel :shrug::skin-tone-6:.
The camera mounting holes are in the edge cuts. Should these be through holes only? I am figuring we mount the camera on top of the pcb, and put screws through it, instead of cutting out a large rectangle for the camera to mount straight to the aluminum.
As I wrote that, I realized we will need places for the bolt heads on the other side.
The pcb mounting holes in the corners are also in the edge cuts.